+++ lib/dry/transaction/step_adapters/try.rb
@@ -1,19 +1,20 @@
 module Dry
-  class Transaction
+  module Transaction
     class StepAdapters
       # @api private
       class Try
-        include Dry::Monads::Either::Mixin
+        include Dry::Monads::Result::Mixin
 
         def call(step, input, *args)
           unless step.options[:catch]
             raise ArgumentError, "+try+ steps require one or more exception classes provided via +catch:+"
           end
 
-          Right(step.operation.call(input, *args))
+          result = step.call_operation(input, *args)
+          Success(result)
         rescue *Array(step.options[:catch]) => e
           e = step.options[:raise].new(e.message) if step.options[:raise]
-          Left(e)
+          Failure(e)
         end
       end