+++ lib/dry/transaction/step_failure.rb
@@ -1,13 +1,11 @@
 module Dry
-  class Transaction
+  module Transaction
     # A wrapper for storing together the step that failed
     # and value describing the failure.
     class StepFailure
       attr_reader :step
       attr_reader :value
 
-      # @param step [Step]
-      # @param value [Object]
       def initialize(step, value)
         @step = step
         @value = value