+++ spec/integration/publishing_step_events_spec.rb
@@ -1,24 +1,29 @@
 RSpec.describe "publishing step events" do
-  let(:transaction) {
-    Dry.Transaction(container: container) do
-      map :process
-      step :verify
-      tee :persist
+  let(:container) {
+    Class.new do
+      extend Dry::Container::Mixin
+
+      register :process, -> input { {name: input["name"]} }
+      register :verify,  -> input { input[:name].to_s != "" ? Dry::Monads.Success(input) : Dry::Monads.Failure("no name") }
+      register :persist, -> input { Test::DB << input and true }
     end
   }
 
-  let(:container) {
-    {
-      process:  -> input { {name: input["name"]} },
-      verify:   -> input { input[:name].to_s != "" ? Right(input) : Left("no name") },
-      persist:  -> input { Test::DB << input and true }
-    }
+  let(:transaction) {
+    Class.new do
+      include Dry::Transaction(container: Test::Container)
+
+      map :process, with: :process
+      step :verify, with: :verify
+      tee :persist, with: :persist
+    end.new
   }
 
   let(:subscriber) { spy(:subscriber) }
 
   before do
     Test::DB = []
+    Test::Container = container
   end
 
   context "subscribing to all step events" do
@@ -60,6 +65,36 @@ RSpec.describe "publishing step events" do
       transaction.call("name" => "")
 
       expect(subscriber).to have_received(:verify_failure).with({name: ""}, "no name")
+    end
+  end
+
+  context "subscribing to step events when passing step arguments" do
+    before do
+      transaction.subscribe(verify: subscriber)
+    end
+
+    let(:container) {
+      Class.new do
+        extend Dry::Container::Mixin
+
+        register :process, -> input { {name: input["name"]} }
+        register :verify,  -> input, name { input[:name].to_s == name ? Dry::Monads.Success(input) : Dry::Monads.Failure("wrong name") }
+        register :persist, -> input { Test::DB << input and true }
+      end
+    }
+
+    specify "subscriber receives success event for the specified step" do
+      transaction.with_step_args(verify: ["Jane"]).call("name" => "Jane")
+
+      expect(subscriber).to have_received(:verify_success).with(name: "Jane")
+      expect(subscriber).not_to have_received(:process_success)
+      expect(subscriber).not_to have_received(:persist_success)
+    end
+
+    specify "subscriber receives failure event for the specified step" do
+      transaction.with_step_args(verify: ["Jade"]).call("name" => "")
+
+      expect(subscriber).to have_received(:verify_failure).with({name: ""}, "Jade", "wrong name")
     end
   end
 end