+++ spec/spec_helper.rb
@@ -1,6 +1,14 @@
-if RUBY_ENGINE == "ruby" && RUBY_VERSION >= "2.3"
-  require "simplecov"
-  SimpleCov.start
+if RUBY_ENGINE == 'ruby' && ENV['COVERAGE'] == 'true'
+  require 'yaml'
+  rubies = YAML.load(File.read(File.join(__dir__, '..', '.travis.yml')))['rvm']
+  latest_mri = rubies.select { |v| v =~ /\A\d+\.\d+.\d+\z/ }.max
+
+  if RUBY_VERSION == latest_mri
+    require 'simplecov'
+    SimpleCov.start do
+      add_filter '/spec/'
+    end
+  end
 end
 
 begin
@@ -8,6 +16,9 @@ begin
 rescue LoadError; end
 
 require "dry-transaction"
+require "dry-matcher"
+require "dry-monads"
+require "dry-container"
 
 # Requires supporting ruby files with custom matchers and macros, etc, in
 # spec/support/ and its subdirectories. Files matching `spec/**/*_spec.rb` are