+++ spec/unit/step_spec.rb
@@ -18,7 +18,7 @@ RSpec.describe Dry::Transaction::Step do
     subject { step.call(input) }
 
     context "when operation succeeds" do
-      let(:operation) { proc { |input| Dry::Monads::Either::Right.new(input) } }
+      let(:operation) { proc { |input| Dry::Monads::Result::Success.new(input) } }
 
       it { is_expected.to be_right }
 
@@ -40,14 +40,14 @@ RSpec.describe Dry::Transaction::Step do
     end
 
     context "when operation fails" do
-      let(:operation) { proc { |input| Dry::Monads::Either::Left.new("error") } }
+      let(:operation) { proc { |input| Dry::Monads::Result::Failure.new("error") } }
 
       it { is_expected.to be_left }
 
       it "wraps value in StepFailure" do
         aggregate_failures do
-          expect(subject.value).to be_a Dry::Transaction::StepFailure
-          expect(subject.value.value).to eq "error"
+          expect(subject.left).to be_a Dry::Transaction::StepFailure
+          expect(subject.left.value).to eq "error"
         end
       end
 
@@ -55,6 +55,35 @@ RSpec.describe Dry::Transaction::Step do
         expect(listener).to receive(:test_failure).with(input, "error")
         step.subscribe(listener)
         subject
+      end
+    end
+  end
+
+  describe "#with" do
+    let(:operation) { proc { |a, b| a + b } }
+    context "without arguments" do
+      it "returns itself" do
+        expect(step.with).to eq step
+      end
+    end
+
+    context "with operation argument" do
+      it "returns new instance with only operation changed" do
+        new_operation = proc { |a,b| a * b }
+        new_step = step.with(operation: new_operation)
+        expect(new_step).to_not eq step
+        expect(new_step.operation_name).to eq step.operation_name
+        expect(new_step.operation).to_not eq step.operation
+      end
+    end
+
+    context "with call_args argument" do
+      let(:call_args) { [12] }
+      it "returns new instance with only call_args changed" do
+        new_step = step.with(call_args: call_args)
+        expect(new_step).to_not eq step
+        expect(new_step.operation_name).to eq step.operation_name
+        expect(new_step.call_args).to_not eq step.call_args
       end
     end
   end